VendingMachine/requirements.txt
opencv-python~=4.8.0.74
requests~=2.31.0
torchvision~=0.15.2
torch~=2.0.1
Pillow~=10.0.0